requirements.txt
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
pip==24.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.1.1
setuptools==68.2.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
wheel==0.41.2
SQLAlchemy==2.0.37
pydantic-settings==2.7.1
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
psycopg2==2.9.10
